# Backend/swagger.yaml
openapi: 3.0.0
info:
  title: Pet Shop API
  description: เอกสาร API สำหรับโปรเจกต์ Petshop-Project-CSI400 (อัปเดตตาม Schema)
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: เซิร์ฟเวอร์สำหรับพัฒนา (Local)
...
paths:
...
  /api/addresses:
    get:
      tags:
        - User
      summary: ดึงที่อยู่ทั้งหมดของผู้ใช้
      security:
        - bearerAuth: []
      responses:
        '200':
          description: รายการที่อยู่
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags:
        - User
      summary: เพิ่มที่อยู่ใหม่
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAddressPayload'
            examples:
              default:
                value:
                  recipientName: "สมศรี รักสัตว์"
                  phone: "0898765432"
                  addressLine1: "555/99 อาคาร B"
                  addressLine2: "คอนโดแมวเหมียว"
                  subdistrict: "คลองจั่น"
                  district: "บางกะปิ"
                  province: "กรุงเทพมหานคร"
                  postalCode: "10240"
                  isDefault: true
      responses:
        '201':
          description: สร้างที่อยู่สำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          description: ข้อมูลไม่ถูกต้อง
...
  /api/products:
    get:
      tags:
        - Products
      summary: ดึงรายการสินค้าทั้งหมด (สำหรับลูกค้า)
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, food, toys, accessories]
            example: "food"
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, rating, reviews, price_asc, price_desc]
            example: "rating"
      responses:
        '200':
          description: รายการสินค้า
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/search:
    get:
      tags:
        - Products
      summary: ค้นหาสินค้า
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            example: "อาหารแมว"
      responses:
        '200':
          description: ผลการค้นหา
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: ดึงข้อมูลสินค้าชิ้นเดียว
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 11
      responses:
        '200':
          description: ข้อมูลสินค้า
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: ไม่พบสินค้า
...
  /cart:
    get:
      tags:
        - Cart
      summary: ดึงข้อมูลตะกร้าสินค้า
      security:
        - bearerAuth: []
      responses:
        '200':
          description: รายการสินค้าในตะกร้า
          content:
            application/json:
              schema:
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
...
  /api/wishlists:
    get:
      tags:
        - Wishlist
      summary: ดึงรายการที่ถูกใจ
      security:
        - bearerAuth: []
      responses:
        '200':
          description: รายการที่ถูกใจ
          content:
            application/json:
              schema:
                properties:
                  wishlists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product' # ส่งกลับเป็นข้อมูล Product
    post:
      tags:
        - Wishlist
      summary: เพิ่มรายการที่ถูกใจ
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                product_id:
                  type: integer
              examples:
                default:
                  value:
                    product_id: 12
      responses:
        '201':
          description: เพิ่มสำเร็จ
    delete:
      tags:
        - Wishlist
      summary: ล้างรายการที่ถูกใจทั้งหมด
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ล้างสำเร็จ
...
  /api/orders:
    get:
      tags:
        - Orders
      summary: ดึงคำสั่งซื้อ (ของผู้ใช้)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: รายการคำสั่งซื้อ
          content:
            application/json:
              schema:
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
    post:
      tags:
        - Orders
      summary: สร้างคำสั่งซื้อใหม่
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrderPayload'
            examples:
              default:
                value:
                  items:
                    - product_id: 11
                      product_name: "Nano แชมพูนาโน"
                      product_price: 220
                      quantity: 1
                      subtotal: 220
                    - product_id: 12
                      product_name: "ไม้แมว หนูแขวน"
                      product_price: 70
                      quantity: 2
                      subtotal: 140
                  shipping:
                    name: "สมชาย ใจดี"
                    phone: "0812345678"
                    line1: "123/45 ซอยเพชรเกษม 1"
                    subdistrict: "บางหว้า"
                    district: "ภาษีเจริญ"
                    province: "กรุงเทพมหานคร"
                    postal_code: "10160"
                  courierMethod: "standard"
                  subtotal: 360
                  shipping_fee: 0
                  total: 360
      responses:
        '201':
          description: สร้างคำสั่งซื้อสำเร็จ
          content:
            application/json:
              schema:
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
...
  /api/payments:
    post:
      tags:
        - Payments
      summary: สร้างการชำระเงิน (Pending)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                order_id:
                  type: integer
                amount:
                  type: number
                payment_method:
                  type: string
                  enum: [mock, credit_card, bank_transfer, cash_on_delivery]
              examples:
                default:
                  value:
                    order_id: 58
                    amount: 360.00
                    payment_method: "mock"
      responses:
        '200':
          description: สร้าง Payment (Pending) สำเร็จ
          content:
            application/json:
              schema:
                properties:
                  payment:
                    $ref: '#/components/schemas/Payment'
        '400':
          description: จำนวนเงินไม่ตรง
        '404':
          description: ไม่พบออเดอร์
...
  /api/admin/users:
    get:
      tags:
        - Admin - Users
      summary: (Admin) ดึงผู้ใช้ทั้งหมด
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
            example: "customer"
        - name: role
          in: query
          schema:
            type: string
            enum: ['', customer, admin]
            example: "customer"
      responses:
        '200':
          description: รายการผู้ใช้
        '403':
          description: ไม่มีสิทธิ์ (Admin เท่านั้น)
    post:
      tags:
        - Admin - Users
      summary: (Admin) สร้างผู้ใช้ใหม่
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminNewUserPayload'
            examples:
              newAdmin:
                value:
                  email: "new.admin@example.com"
                  password: "password123"
                  firstName: "แอดมิน"
                  lastName: "ใหม่"
                  phone: "0899999999"
                  role: "admin"
      responses:
        '201':
          description: สร้างผู้ใช้สำเร็จ
        '400':
          description: อีเมลซ้ำ หรือ ข้อมูลไม่ถูกต้อง
        '403':
          description: ไม่มีสิทธิ์
...
  /api/admin/products:
    get:
      tags:
        - Admin - Products
      summary: (Admin) ดึงสินค้าทั้งหมด (สำหรับ Admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: รายการสินค้า
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '403':
          description: ไม่มีสิทธิ์
    post:
      tags:
        - Admin - Products
      summary: (Admin) เพิ่มสินค้าใหม่
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "อาหารแมว Kaniva"
                description:
                  type: string
                  example: "สูตรเนื้อไก่ ปลาทูน่าและข้าว สำหรับแมวทุกวัย"
                category:
                  type: string
                  enum: [food, toys, accessories]
                  example: "food"
                price:
                  type: number
                  example: 199
                stock:
                  type: integer
                  example: 100
                newImages:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: เพิ่มสินค้าสำเร็จ
        '403':
          description: ไม่มีสิทธิ์
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "ใส่ Access Token (JWT) ที่ได้จากการ Login (ตัวอย่าง: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...) **ไม่ต้องใส่ 'Bearer' นำหน้า**"
  schemas:
...
    NewAddressPayload:
      type: object
      description: ข้อมูลสำหรับสร้างหรืออัปเดตที่อยู่
      properties:
        recipientName:
          type: string
          example: "สมศรี รักสัตว์"
        phone:
          type: string
          example: "0898765432"
        addressLine1:
          type: string
          example: "555/99 อาคาร B"
        addressLine2:
          type: string
          example: "คอนโดแมวเหมียว"
        subdistrict:
          type: string
          example: "คลองจั่น"
        district:
          type: string
          example: "บางกะปิ"
        province:
          type: string
          example: "กรุงเทพมหานคร"
        postalCode:
          type: string
          example: "10240"
        isDefault:
          type: boolean
          example: false
      required: [recipientName, phone, addressLine1, subdistrict, district, province, postalCode]
...
    NewOrderPayload:
      type: object
      description: ข้อมูลสำหรับสร้างออเดอร์ใหม่
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
              product_name:
                type: string
              product_price:
                type: number
              quantity:
                type: integer
              subtotal:
                type: number
        shipping:
          type: object
          description: Snapshot ของที่อยู่ตอนสั่งซื้อ
          properties:
            name: { type: string }
            phone: { type: string }
            line1: { type: string }
            subdistrict: { type: string }
            district: { type: string }
            province: { type: string }
            postal_code: { type: string }
        courierMethod:
          type: string
          example: "standard"
        subtotal:
          type: number
          example: 1870
        shipping_fee:
          type: number
          example: 0
        total:
          type: number
          example: 1870
    AdminNewUserPayload:
      type: object
      properties:
        email:
          type: string
          example: "new.admin@example.com"
        password:
          type: string
          example: "password123"
        firstName:
          type: string
          example: "แอดมิน"
        lastName:
          type: string
          example: "ใหม่"
        phone:
          type: string
          example: "0899999999"
        role:
          type: string
          enum: [customer, admin]
          example: "admin"
      required: [email, password, firstName, lastName, role]
...
    Product:
      type: object
      description: โมเดล Product จากตาราง products
      properties:
        id: { type: integer, example: 11 }
        name: { type: string, example: "Nano แชมพูนาโน" }
        description: { type: string, example: "แชมพูสำหรับสัตว์เลี้ยง 280ml." }
        category: { type: string, example: "accessories" }
        price: { type: number, format: "float", example: 220.00 }
        stock: { type: integer, example: 48 }
        images: { type: array, items: { type: string }, example: ["/uploads/nano-shom-1762589997422.jpg"] }
        rating: { type: number, format: "float", example: 0.0 }
        reviews: { type: integer, example: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Address:
      type: object
      description: โมเดล Address จากตาราง addresses
      properties:
        id: { type: integer, example: 1 }
        user_id: { type: integer, example: 1 }
        recipient_name: { type: string, example: "สมชาย ใจดี" }
        phone: { type: string, example: "0812345678" }
        address_line1: { type: string, example: "123/45 ซอยเพชรเกษม 1" }
        address_line2: { type: string, example: "หมู่บ้านสุขใจ" }
        subdistrict: { type: string, example: "บางหว้า" }
        district: { type: string, example: "ภาษีเจริญ" }
        province: { type: string, example: "กรุงเทพมหานคร" }
        postal_code: { type: string, example: "10160" }
        is_default: { type: boolean, example: true }
        is_deleted: { type: boolean, example: false }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CartItem:
      type: object
      description: โมเดล CartItem จากตาราง cart_items
      properties:
        id: { type: integer, description: "product_id", example: 11 }
        name: { type: string, example: "Nano แชมพูนาโน" }
        price: { type: number, example: 220.00 }
        images: { type: array, items: { type: string }, example: ["/uploads/nano-shom-1762589997422.jpg"] }
        quantity: { type: integer, example: 2 }
...
    Order:
      type: object
      description: โมเดล Order จากตาราง orders
      properties:
        id: { type: integer, example: 58 }
        user_id: { type: integer, example: 3 }
        order_number: { type: string, example: "ORD-MHUURI1I" }
        status: { 
          type: string, 
          enum: [pending_payment, paid, preparing, ready_to_ship, shipping, completed, cancelled],
          example: "pending_payment"
        }
        shipping_recipient_name: { type: string, example: "สมชาย ใจดี" }
        shipping_phone: { type: string, example: "0812345678" }
        shipping_address_line1: { type: string, example: "123/45 ซอยเพชรเกษม 1" }
        shipping_address_line2: { type: string, example: "" }
        shipping_subdistrict: { type: string, example: "บางหว้า" }
        shipping_district: { type: string, example: "ภาษีเจริญ" }
        shipping_province: { type: string, example: "กรุงเทพมหานคร" }
        shipping_postal_code: { type: string, example: "10160" }
        courier_method: { type: string, example: "standard" }
        subtotal: { type: number, example: 1870.00 }
        shipping_fee: { type: number, example: 0 }
        total: { type: number, example: 1870.00 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        delivered_at: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
    OrderItem:
      type: object
      description: โมเดล OrderItem จากตาราง order_items
      properties:
        id: { type: integer }
        order_id: { type: integer }
        product_id: { type: integer, example: 11 }
        product_name: { type: string, example: "Nano แชมพูนาโน" }
        product_price: { type: number, example: 220.00 }
        quantity: { type: integer, example: 1 }
        subtotal: { type: number, example: 220.00 }
    Payment:
      type: object
      description: โมเดล Payment จากตาราง payments
      properties:
        id: { type: integer }
        order_id: { type: integer }
        user_id: { type: integer }
        payment_method: { type: string, example: "mock" }
        amount: { type: number, example: 1870.00 }
        payment_status: { 
          type: string, 
          enum: [pending, success, failed, refunded],
          example: "pending"
        }
        transaction_id: { type: string, example: "MOCK-12345" }
        payment_date: { type: string, format: date-time }